Revert "IP detection of locations can occur locally, falling back to LOC attribute."
This reverts commit 7d1524e36a3be791381a1ade3fd1e41f61c45633.

That commit should be rewritten and applied on the adc branch.

We decided to only do DNS lookups for your own "Location", and use IP lookups
to figure out other nodes' "Location"s. This is to reduce DNS lookups over the
entire network. (As a corollary, "Location" claims from other nodes should not
be treated as authoritative; it should be ignored if a IP lookup on the local
side contradicts it.)

git-svn-id: https://dtella-cambridge.googlecode.com/svn/trunk@390 efa14470-2078-11de-9d52-39b120432c5d